% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Clean raw data for analysis}
\usage{
clean_data(x)
}
\arguments{
\item{x}{A dataframe probably from MasterThesisData2024.csv}
}
\value{
A dataframe ready for analysis
}
\description{
Clean raw data for analysis
}
\examples{
\dontrun{
MasterThesisData <- read.table(
file = "data/raw/MasterThesisData2024.csv",
header = TRUE,
sep = ",",
na.strings = "NA",
stringsAsFactors = TRUE,
dec = "."
)
TreMs <- clean_data(MasterThesisData)
}
}
