% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_deadwood_identities.R
\name{create_deadwood_identities}
\alias{create_deadwood_identities}
\title{Deadwood identities - Create a column that combines species and type of deadwood}
\usage{
create_deadwood_identities(TreMs)
}
\arguments{
\item{TreMs}{A TreMs dataframe}
}
\value{
A TreMs dataframe with a DeadwoodIdentities column
}
\description{
Deadwood identities - Create a column that combines species and type of deadwood
}
\examples{
\dontrun{
TreMs <- create_deadwood_identities(TreMs)
}
}
