% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_models.R
\name{create_models}
\alias{create_models}
\title{Create all the models}
\usage{
create_models(TreMs)
}
\arguments{
\item{TreMs}{A TreMs dataframe}
}
\value{
A dataframe containing all the models
}
\description{
Create all the models
}
\examples{
\dontrun{
mod_df <- create_models(TreMs)
}
}
