% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_models.R
\name{get_model_cols}
\alias{get_model_cols}
\title{Make a tibble with model variables and families for create_models}
\usage{
get_model_cols()
}
\value{
A tibble with columns for variable and model_family
}
\description{
Make a tibble with model variables and families for create_models
}
