% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_results.R
\name{get_model_object}
\alias{get_model_object}
\title{Get model object from a dataframe}
\usage{
get_model_object(x, model_index = 1)
}
\arguments{
\item{x}{A dataframe of models from create_models()}

\item{model_index}{Row number from x to get the formula from (optional)}
}
\value{
A model object
}
\description{
Get model object from a dataframe
}
\examples{
\dontrun{
get_model_object(mod_df, 1)
}
}
