% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_results.R
\name{get_residuals}
\alias{get_residuals}
\title{Get model simulated residuals}
\usage{
get_residuals(x, model_index)
}
\arguments{
\item{x}{A dataframe of models from create_models()}

\item{model_index}{Row number from x to get the residuals from}
}
\value{
Model simulated residuals from DHARMa::simulateResiduals
}
\description{
Get model simulated residuals
}
\examples{
\dontrun{
get_residuals(mod_df, 1)
}
}
