% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{group_data}
\alias{group_data}
\title{Group data by species and deadwood type}
\usage{
group_data(TreMs)
}
\arguments{
\item{TreMs}{A dataframe from summarize_microhabitats}
}
\value{
A dataframe
}
\description{
Group data by species and deadwood type
}
\examples{
\dontrun{
  TreMs <- load_data(MasterThesisData2024)
  TreMs <- summarize_microhabitats(TreMs)
  TreMs <- group_data(TreMs)
}
}
