% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jovanovic2025-package.R
\docType{package}
\name{jovanovic2025-package}
\alias{jovanovic2025}
\alias{jovanovic2025-package}
\title{jovanovic2025: Jovanovic 2025 Thesis Code}
\description{
Corresponding package to Jovanovic 2025. Please read the paper.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pandionlabs.github.io/jovanovic2025/}
}

}
\author{
\strong{Maintainer}: Sam Herniman \email{sam@pandionlabs.dev} (\href{https://orcid.org/0000-0001-7663-2897}{ORCID}) [compiler]

Authors:
\itemize{
  \item Magdalena Jovanovic [copyright holder]
  \item Simone Massaro \email{simone@pandionlabs.dev} (\href{https://orcid.org/0009-0005-7459-8935}{ORCID}) [compiler]
}

}
\keyword{internal}
