% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{load_data}
\alias{load_data}
\title{Load the data and prepare it for analysis}
\usage{
load_data(MasterThesisData)
}
\arguments{
\item{MasterThesisData}{A dataframe probably from MasterThesisData2024.csv}
}
\value{
A dataframe of tree data
}
\description{
Load the data and prepare it for analysis
}
\examples{
\dontrun{
MasterThesisData <- read.table(
file = "data/raw/MasterThesisData2024.csv",
header = TRUE,
sep = ",",
na.strings = "NA",
stringsAsFactors = TRUE,
dec = "."
)
  TreMs <- load_data(MasterThesisData)
}
}
