% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MasterThesisData2024}
\alias{MasterThesisData2024}
\title{Data for the Javanovic Masters Thesis}
\format{
\subsection{\code{MasterThesisData2024}}{

A data frame with 534 rows and 89 columns:
\describe{
\item{Plot}{plot number}
\item{Treedata.Treeid}{tree id number}
}
}
}
\source{
IDENTIFYING AND PREDICTING TREE RELATED MICROHABITATS ON DOWNED DEADWOOD AND STUMPS Masters Thesis
}
\usage{
MasterThesisData2024
}
\description{
IDENTIFYING AND PREDICTING TREE RELATED MICROHABITATS ON DOWNED DEADWOOD AND STUMPS
}
\keyword{datasets}
