% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_identities.R
\name{summarize_identities}
\alias{summarize_identities}
\title{Create a summary table for TreMs data}
\usage{
summarize_identities(x, grouping_variable)
}
\arguments{
\item{x}{TreMs dataframe}

\item{grouping_variable}{variable to be used for grouping}
}
\value{
A summary table
}
\description{
Create a summary table for TreMs data
}
\examples{
\dontrun{
TreeIdentitiesSummary <- summarize_identities(
TreMs, 
grouping_variable = TreeIdentities2
)
}
}
