% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{summarize_microhabitats}
\alias{summarize_microhabitats}
\title{Summarizes microhabitats in the TreMs dataset}
\usage{
summarize_microhabitats(TreMs)
}
\arguments{
\item{TreMs}{a dataframe from load_data}
}
\value{
A dataframe of summarized microhabitat data
}
\description{
Combines single type of microhabitat observations into larger groups
}
\examples{
\dontrun{
  TreMs <- load_data(MasterThesisData2024)
  TreMs <- summarize_microhabitats(TreMs)
}
}
