% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_models.R
\name{create_model}
\alias{create_model}
\title{Build a model and return fit statistics}
\usage{
create_model(variable, model_family, TreMs)
}
\arguments{
\item{variable}{Character specifying the y variable for the model}

\item{model_family}{Model family passed to glmmTMB::glmmTMB}

\item{TreMs}{A TreMs dataframe}
}
\value{
A tibble containing a model and statistics
}
\description{
Build a model and return fit statistics
}
\details{
Builds a glmmTMB::glmmTMB model using the formula
\{variable\} ~ GroupedTreeSpecies + Treedata.DBH_cm + Treedata.Tree_Decay + (1|Plot)
Where the Y variable is selected from the available columns in TreMs
}
\examples{
\dontrun{
TreMs <- clean_data(MasterThesisData2024)
richness <- create_model("Richness", "poisson", TreMs)
}
}
