% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_results.R
\name{get_formula}
\alias{get_formula}
\title{Get model formula}
\usage{
get_formula(x, model_index)
}
\arguments{
\item{x}{A dataframe of models from create_models()}

\item{model_index}{Row number from x to get the formula from}
}
\value{
A model formula
}
\description{
Get model formula
}
\examples{
\dontrun{
get_formula(mod_df, 1)
}
}
