% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_results.R
\name{get_test_dispersion}
\alias{get_test_dispersion}
\title{Test a model for dispersion}
\usage{
get_test_dispersion(x, model_index)
}
\arguments{
\item{x}{A dataframe of models from create_models()}

\item{model_index}{Row number from x to get the dispersion test result from}
}
\value{
Result of DHARMa::testDispersion
}
\description{
Test a model for dispersion
}
\examples{
\dontrun{
get_test_dispersion(mod_df, 1)
}
}
