% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_results.R
\name{get_test_outliers}
\alias{get_test_outliers}
\title{Test a model for outliers}
\usage{
get_test_outliers(x, model_index = 1)
}
\arguments{
\item{x}{A dataframe of models from create_models()}

\item{model_index}{Row number from x to get the outliers from}
}
\value{
Result of DHARMa::testOutliers
}
\description{
Test a model for outliers
}
\examples{
\dontrun{
get_residuals(mod_df, 1)
}
}
