% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_results.R
\name{get_test_zero_inflation}
\alias{get_test_zero_inflation}
\title{Test a model for zero inflation}
\usage{
get_test_zero_inflation(x, model_index)
}
\arguments{
\item{x}{A dataframe of models from create_models()}

\item{model_index}{Row number from x to get the zero inflation test result from}
}
\value{
Result of DHARMa::testZeroInflation
}
\description{
Test a model for zero inflation
}
\examples{
\dontrun{
get_test_zero_inflation(mod_df, 1)
}
}
