% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{group_data}
\alias{group_data}
\title{Group data by species and deadwood type}
\usage{
group_data(TreMs)
}
\arguments{
\item{TreMs}{A dataframe from summarize_microhabitats}
}
\value{
A dataframe
}
\description{
Group data by species and deadwood type
}
\examples{
\dontrun{
MasterThesisData <- read.table(
file = "data/raw/MasterThesisData2024.csv",
header = TRUE,
sep = ",",
na.strings = "NA",
stringsAsFactors = TRUE,
dec = "."
)
  TreMs <- load_data(MasterThesisData)
  TreMs <- summarize_microhabitats(TreMs)
  TreMs <- group_data(TreMs)
}
}
