% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_count_table_by_decay.R
\name{make_count_table_by_decay}
\alias{make_count_table_by_decay}
\title{Make a count table by decay stage}
\usage{
make_count_table_by_decay(
  TreMs,
  save_name = here::here("data/derivatives/Count_Table_by_Decay_Stage.csv")
)
}
\arguments{
\item{TreMs}{A TreMs dataframe}

\item{save_name}{Path to save the dataframe}
}
\value{
A count table dataframe
}
\description{
Make a count table by decay stage
}
\examples{
\dontrun{
make_count_table_by_decay(TreMs, save_name = "Count_Table_by_Decay_Stage.csv")
}
}
