% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{summarize_microhabitats}
\alias{summarize_microhabitats}
\title{Summarizes microhabitats in the TreMs dataset}
\usage{
summarize_microhabitats(TreMs)
}
\arguments{
\item{TreMs}{a dataframe from load_data}
}
\value{
A dataframe of summarized microhabitat data
}
\description{
Combines single type of microhabitat observations into larger groups
}
\examples{
\dontrun{
MasterThesisData <- read.table(
file = "data/raw/MasterThesisData2024.csv",
header = TRUE,
sep = ",",
na.strings = "NA",
stringsAsFactors = TRUE,
dec = "."
)
  TreMs <- load_data(MasterThesisData)
  TreMs <- summarize_microhabitats(TreMs)
}
}
